Create gapfilling module
Move function create_extended_model() from fastgapfill to the
new gapfilling module. Take methods from MetabolicModel related
to building the extended model and move them as functions into
the new gapfilling module.
