+++ psamm/metabolicmodel.py
@@ -268,112 +268,6 @@ class MetabolicModel(MetabolicDatabase):
                    for other_reaction in reactions):
                 self._compound_set.remove(compound)
 
-    def add_all_database_reactions(self, compartments):
-        """Add all reactions from database that occur in given compartments"""
-
-        added = set()
-        for rxnid in self._database.reactions:
-            reaction = self._database.get_reaction(rxnid)
-            if all(compound.compartment in compartments
-                   for compound, _ in reaction.compounds):
-                if rxnid not in self._reaction_set:
-                    added.add(rxnid)
-                self.add_reaction(rxnid)
-
-        return added
-
-    def add_all_exchange_reactions(self, compartment, allow_duplicates=False):
-        """Add all exchange reactions to database and to model"""
-
-        all_reactions = {}
-        if not allow_duplicates:
-            # TODO: Avoid adding reactions that already exist in the database.
-            # This should be integrated in the database.
-            for rxnid in self._database.reactions:
-                rx = self._database.get_reaction(rxnid)
-                all_reactions[rx] = rxnid
-
-        added = set()
-        added_compounds = set()
-        initial_compounds = set(self.compounds)
-        for model_compound in initial_compounds:
-            compound = model_compound.in_compartment(compartment)
-            if compound in added_compounds:
-                continue
-
-            rxnid_ex = create_exchange_id(self._database.reactions, compound)
-
-            reaction_ex = Reaction(Direction.Both, {compound: -1})
-            if reaction_ex not in all_reactions:
-                self._database.set_reaction(rxnid_ex, reaction_ex)
-            else:
-                rxnid_ex = all_reactions[reaction_ex]
-
-            if rxnid_ex not in self._reaction_set:
-                added.add(rxnid_ex)
-            self.add_reaction(rxnid_ex)
-            added_compounds.add(compound)
-
-        return added
-
-    def add_all_transport_reactions(self, boundaries, allow_duplicates=False):
-        """Add all transport reactions to database and to model.
-
-        Add transport reactions for all boundaries. Boundaries are defined
-        by pairs (2-tuples) of compartment IDs. Transport reactions are
-        added for all compounds in the model, not just for compounds in the
-        two boundary compartments.
-
-        Args:
-            boundaries: Set of compartment boundary pairs.
-
-        Returns:
-            Set of IDs of reactions that were added.
-        """
-
-        all_reactions = {}
-        if not allow_duplicates:
-            # TODO: Avoid adding reactions that already exist in the database.
-            # This should be integrated in the database.
-            for rxnid in self._database.reactions:
-                rx = self._database.get_reaction(rxnid)
-                all_reactions[rx] = rxnid
-
-        boundary_pairs = set()
-        for source, dest in boundaries:
-            if source != dest:
-                boundary_pairs.add(tuple(sorted((source, dest))))
-
-        added = set()
-        added_pairs = set()
-        initial_compounds = set(self.compounds)
-        for compound in initial_compounds:
-            for c1, c2 in boundary_pairs:
-                compound1 = compound.in_compartment(c1)
-                compound2 = compound.in_compartment(c2)
-                pair = compound1, compound2
-                if pair in added_pairs:
-                    continue
-
-                rxnid_tp = create_transport_id(
-                    self._database.reactions, compound1, compound2)
-
-                reaction_tp = Reaction(Direction.Both, {
-                    compound1: -1,
-                    compound2: 1
-                })
-                if reaction_tp not in all_reactions:
-                    self._database.set_reaction(rxnid_tp, reaction_tp)
-                else:
-                    rxnid_tp = all_reactions[reaction_tp]
-
-                if rxnid_tp not in self._reaction_set:
-                    added.add(rxnid_tp)
-                self.add_reaction(rxnid_tp)
-                added_pairs.add(pair)
-
-        return added
-
     def copy(self):
         """Return copy of model"""
